regs package has moved to internal
