+++ internal/pkg/regs/regs_math.go
